+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-<!-- TODO: describe breaking changes and new features -->
+<!-- TODO: describe breaking changes and new features, don't forget to write a migration guide -->
 
 ## 1.4.0